pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "pydantic==2.12.5",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests>=2.32.5",
    "uvicorn==0.24.0",
]
